iio.h: Add API function iio_block_disable_cpu_access()
This function is meant to be used when the data is never accessed
directly, but passed around as DMABUF objects. In that case, keeping the
data in DMA space means that it doesn't have to be synchronized with the
CPU cache, which improves performance, and saves two ioctl calls.

When CPU access is disabled, all API functions to access a block's data
should not be used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>